refactor(components): group homepage components and extract shared UI
Move Hero, DatasetOverview, UbMojiDatasetSection/Tables, AuthorSection
and CitationSection into components/home/ so components are grouped by
feature. Rename NewsLayout to PageLayout since it is also used by the
papers pages and 404 (and drop its unused locale prop). Extract the
duplicated section shell into HomeSection.astro and the duplicated
empty-state card into EmptyState.astro, reuse DatasetTableCard for the
naming-convention table, and replace `Astro.props as Props` casts with
typed Astro.props. Delete the stale hero README.
