+++ src/components/EmptyState.astro
@@ -1,0 +1,15 @@
+---
+type Props = {
+  title: string;
+  body: string;
+};
+
+const { title, body } = Astro.props;
+---
+
+<div class="rounded-3xl border border-dashed border-border bg-muted/30 p-8 text-center sm:p-12">
+  <div class="mx-auto max-w-2xl space-y-3">
+    <h2 class="text-2xl font-semibold tracking-tight">{title}</h2>
+    <p class="text-sm leading-7 text-muted-foreground sm:text-base">{body}</p>
+  </div>
+</div>
