+++ src/components/Footer.astro
@@ -1,10 +1,14 @@
 ---
 import { Image } from "astro:assets";
-import { defaultLocale, isLocale, t } from "@/lib/i18n";
+import { resolveLocale, t } from "@/lib/i18n";
 import logoSrc from "@/assets/logo.png";
 
+type Props = {
+  lang?: string;
+};
+
 const { lang } = Astro.props;
-const locale = isLocale(lang) ? lang : defaultLocale;
+const locale = resolveLocale(lang);
 ---
 
 <footer data-testid="site-footer" class="border-t border-border/60 bg-muted/40">
