--- src/components/dataset/UbMojiDatasetSection.astro
@@ -1,48 +1,0 @@
----
-import AuthorSection from "@/components/AuthorSection.astro";
-import { authors } from "@/lib/authors";
-import type { Locale } from "@/lib/i18n";
-import { t } from "@/lib/i18n";
-import { getUbMojiDatasetDetails } from "@/lib/ub-moji-dataset-details";
-import UbMojiDatasetTables from "@/components/dataset/UbMojiDatasetTables";
-import CitationSection from "@/components/CitationSection.astro";
-
-type Props = { locale: Locale };
-
-const { locale } = Astro.props as Props;
-const details = getUbMojiDatasetDetails(locale);
----
-
-<section class="w-full py-16 sm:py-20" data-testid="dataset-details">
-  <div class="mx-auto w-full max-w-5xl px-4">
-    <div class="border-t border-border/60 pt-10 sm:pt-12">
-      <div class="max-w-3xl">
-        <p
-          class="text-balance text-2xl font-semibold tracking-[0.02em] text-foreground sm:text-3xl md:text-4xl"
-        >
-          {details.title}
-        </p>
-        <p class="mt-4 text-base leading-relaxed text-muted-foreground sm:text-lg">
-          {details.intro}
-        </p>
-        <p class="mt-3 text-sm text-muted-foreground">{details.consentNote}</p>
-      </div>
-
-      <div class="mt-10">
-        <UbMojiDatasetTables details={details} />
-      </div>
-
-      <div class="mt-16">
-        <AuthorSection
-          title={t(locale, "authorSectionTitle")}
-          authors={authors}
-          affiliation={t(locale, "authorSectionAffiliation")}
-        />
-      </div>
-
-      <div class="mt-16">
-        <CitationSection locale={locale} />
-      </div>
-    </div>
-  </div>
-</section>
