+++ src/components/home/AuthorSection.astro
@@ -8,7 +8,7 @@ type Props = {
   className?: string;
 };
 
-const { title, authors, affiliation, className } = Astro.props as Props;
+const { title, authors, affiliation, className } = Astro.props;
 ---
 
 <div class={`flex flex-col gap-8 ${className ?? ""}`}>
