+++ src/components/home/DatasetOverview.astro
@@ -1,0 +1,36 @@
+---
+import DatasetBadge from "@/assets/dataset-on-hf-lg.svg";
+import DatasetBadgeDark from "@/assets/dataset-on-hf-lg-dark.svg";
+import HomeSection from "@/components/home/HomeSection.astro";
+import type { Locale } from "@/lib/i18n";
+import { t } from "@/lib/i18n";
+
+type Props = {
+  locale: Locale;
+};
+
+const { locale } = Astro.props;
+---
+
+<HomeSection
+  testId="dataset-overview"
+  title={t(locale, "datasetTitle")}
+  intro={t(locale, "datasetBody")}
+>
+  <div class="mt-8 flex flex-wrap items-center gap-4 sm:mt-10">
+    <a
+      class="group inline-flex items-center p-1 transition-transform duration-300 hover:-translate-y-0.5 focus-visible:outline-none focus-visible:ring-4 focus-visible:ring-ring/30"
+      href="https://huggingface.co/datasets/kanglabs/ub-MOJI"
+      aria-label={t(locale, "datasetCtaAlt")}
+    >
+      <DatasetBadge
+        class="h-12 w-auto drop-shadow-[0_8px_24px_rgba(0,0,0,0.12)] sm:h-14 dark:hidden"
+        aria-hidden="true"
+      />
+      <DatasetBadgeDark
+        class="hidden h-12 w-auto drop-shadow-[0_8px_24px_rgba(0,0,0,0.35)] sm:h-14 dark:block"
+        aria-hidden="true"
+      />
+    </a>
+  </div>
+</HomeSection>
