+++ src/components/home/UbMojiDatasetSection.astro
@@ -1,0 +1,38 @@
+---
+import AuthorSection from "@/components/home/AuthorSection.astro";
+import CitationSection from "@/components/home/CitationSection.astro";
+import HomeSection from "@/components/home/HomeSection.astro";
+import UbMojiDatasetTables from "@/components/home/UbMojiDatasetTables";
+import { authors } from "@/lib/authors";
+import type { Locale } from "@/lib/i18n";
+import { t } from "@/lib/i18n";
+import { getUbMojiDatasetDetails } from "@/lib/ub-moji-dataset-details";
+
+type Props = { locale: Locale };
+
+const { locale } = Astro.props;
+const details = getUbMojiDatasetDetails(locale);
+---
+
+<HomeSection
+  testId="dataset-details"
+  title={details.title}
+  intro={details.intro}
+  note={details.consentNote}
+>
+  <div class="mt-10">
+    <UbMojiDatasetTables details={details} />
+  </div>
+
+  <div class="mt-16">
+    <AuthorSection
+      title={t(locale, "authorSectionTitle")}
+      authors={authors}
+      affiliation={t(locale, "authorSectionAffiliation")}
+    />
+  </div>
+
+  <div class="mt-16">
+    <CitationSection locale={locale} />
+  </div>
+</HomeSection>
