+++ src/components/news/NewsArticle.astro
@@ -12,8 +12,7 @@ type Props = {
   title: string;
 };
 
-const { backHref, locale, publishedAt, summary, tag, title } =
-  Astro.props as Props;
+const { backHref, locale, publishedAt, summary, tag, title } = Astro.props;
 
 const formatter = new Intl.DateTimeFormat(locale, { dateStyle: "medium" });
 ---
