+++ src/components/news/NewsIndex.astro
@@ -1,4 +1,5 @@
 ---
+import EmptyState from "@/components/EmptyState.astro";
 import type { NewsIndexItem } from "@/lib/news";
 import type { Locale } from "@/lib/i18n";
 import { t } from "@/lib/i18n";
@@ -8,7 +9,7 @@ type Props = {
   locale: Locale;
 };
 
-const { entries, locale } = Astro.props as Props;
+const { entries, locale } = Astro.props;
 
 const formatter = new Intl.DateTimeFormat(locale, {
   dateStyle: "medium",
@@ -51,13 +52,9 @@ const formatter = new Intl.DateTimeFormat(locale, {
       ))}
     </div>
   ) : (
-    <div class="rounded-3xl border border-dashed border-border bg-muted/30 p-8 text-center sm:p-12">
-      <div class="mx-auto max-w-2xl space-y-3">
-        <h2 class="text-2xl font-semibold tracking-tight">{t(locale, "newsEmptyTitle")}</h2>
-        <p class="text-sm leading-7 text-muted-foreground sm:text-base">
-          {t(locale, "newsEmptyBody")}
-        </p>
-      </div>
-    </div>
+    <EmptyState
+      title={t(locale, "newsEmptyTitle")}
+      body={t(locale, "newsEmptyBody")}
+    />
   )
 }
