+++ src/components/news/NewsNotFoundState.astro
@@ -7,7 +7,7 @@ type Props = {
   locale: Locale;
 };
 
-const { indexHref, locale } = Astro.props as Props;
+const { indexHref, locale } = Astro.props;
 ---
 
 <div class="rounded-[2rem] border border-dashed border-border bg-muted/30 p-8 sm:p-10">
