+++ src/components/news/NewsUnavailableState.astro
@@ -8,7 +8,7 @@ type Props = {
   locale: Locale;
 };
 
-const { defaultArticleHref, indexHref, locale } = Astro.props as Props;
+const { defaultArticleHref, indexHref, locale } = Astro.props;
 ---
 
 <div class="rounded-[2rem] border border-border/70 bg-card/80 p-8 shadow-xs sm:p-10">
