+++ src/components/papers/PapersSection.astro
@@ -1,5 +1,6 @@
 ---
 import type { CollectionEntry } from "astro:content";
+import EmptyState from "@/components/EmptyState.astro";
 import type { Locale } from "@/lib/i18n";
 import { t } from "@/lib/i18n";
 
@@ -8,7 +9,7 @@ type Props = {
   locale: Locale;
 };
 
-const { papers, locale } = Astro.props as Props;
+const { papers, locale } = Astro.props;
 const sorted = [...papers].sort((a, b) => a.data.order - b.data.order);
 ---
 
@@ -54,15 +55,9 @@ const sorted = [...papers].sort((a, b) => a.data.order - b.data.order);
       ))}
     </div>
   ) : (
-    <div class="rounded-3xl border border-dashed border-border bg-muted/30 p-8 text-center sm:p-12">
-      <div class="mx-auto max-w-2xl space-y-3">
-        <h2 class="text-2xl font-semibold tracking-tight">
-          {t(locale, "papersEmptyTitle")}
-        </h2>
-        <p class="text-sm leading-7 text-muted-foreground sm:text-base">
-          {t(locale, "papersEmptyBody")}
-        </p>
-      </div>
-    </div>
+    <EmptyState
+      title={t(locale, "papersEmptyTitle")}
+      body={t(locale, "papersEmptyBody")}
+    />
   )
 }
