+++ src/layouts/main.astro
@@ -3,7 +3,8 @@ import { SEO } from "astro-seo";
 import "@/styles/global.css";
 import Header from "@/components/Header";
 import Footer from "@/components/Footer.astro";
-import { defaultLocale, isLocale, t } from "@/lib/i18n";
+import { resolveLocale, t } from "@/lib/i18n";
+import { withTrailingSlash } from "@/lib/routes";
 
 interface Props {
   lang?: string;
@@ -23,11 +24,12 @@ const {
   nofollow = false,
   ogType = "website",
   ogImage,
-} = Astro.props as Props;
-const locale = lang !== undefined && isLocale(lang) ? lang : defaultLocale;
+} = Astro.props;
+const locale = resolveLocale(lang);
 const pathname = Astro.url.pathname;
 const basePath = import.meta.env.BASE_URL;
-const faviconHref = `${basePath.replace(/\/?$/, "/")}favicon.ico`;
+const normalizedBase = withTrailingSlash(basePath);
+const faviconHref = `${normalizedBase}favicon.ico`;
 
 const siteTitle = "ub-MOJI";
 const siteFullTitle = "ub-MOJI | AI Vision Lab";
@@ -37,7 +39,6 @@ const canonicalURL = Astro.site
   ? new URL(Astro.url.pathname, Astro.site).href
   : Astro.url.href;
 const ogLocale = locale === "ja" ? "ja_JP" : "en_US";
-const normalizedBase = basePath.replace(/\/?$/, "/");
 const defaultOgImageUrl = Astro.site
   ? new URL(`${normalizedBase}og/default.png`, Astro.site).href
   : `${normalizedBase}og/default.png`;
