+++ src/pages/404.astro
@@ -1,5 +1,5 @@
 ---
-import NewsLayout from "@/components/news/NewsLayout.astro";
+import PageLayout from "@/components/PageLayout.astro";
 import NewsNotFoundState from "@/components/news/NewsNotFoundState.astro";
 import Layout from "@/layouts/main.astro";
 import { getLocaleFromPath, t } from "@/lib/i18n";
@@ -10,8 +10,7 @@ const basePath = import.meta.env.BASE_URL;
 ---
 
 <Layout lang={locale} noindex={true}>
-  <NewsLayout
-    locale={locale}
+  <PageLayout
     eyebrow="404"
     title={t(locale, "newsNotFoundTitle")}
     intro={t(locale, "newsNotFoundBody")}
@@ -21,5 +20,5 @@ const basePath = import.meta.env.BASE_URL;
       indexHref={buildSectionIndexPath("news", locale, basePath)}
       locale={locale}
     />
-  </NewsLayout>
+  </PageLayout>
 </Layout>
