+++ src/pages/[...lang]/index.astro
@@ -1,7 +1,7 @@
 ---
-import DatasetOverview from "@/components/dataset/DatasetOverview.astro";
-import UbMojiDatasetSection from "@/components/dataset/UbMojiDatasetSection.astro";
-import Hero from "@/components/hero/Hero.astro";
+import DatasetOverview from "@/components/home/DatasetOverview.astro";
+import Hero from "@/components/home/Hero.astro";
+import UbMojiDatasetSection from "@/components/home/UbMojiDatasetSection.astro";
 import Layout from "@/layouts/main.astro";
 import { resolveLocale } from "@/lib/i18n";
 import { getLangStaticPaths } from "@/lib/routes";
