+++ src/pages/[...lang]/papers/index.astro
@@ -1,5 +1,5 @@
 ---
-import NewsLayout from "@/components/news/NewsLayout.astro";
+import PageLayout from "@/components/PageLayout.astro";
 import PapersSection from "@/components/papers/PapersSection.astro";
 import Layout from "@/layouts/main.astro";
 import { resolveLocale, t } from "@/lib/i18n";
@@ -17,12 +17,11 @@ const papers = await getCollection("papers");
   title={t(locale, "papersIndexTitle")}
   description={t(locale, "papersIndexIntro")}
 >
-  <NewsLayout
-    locale={locale}
+  <PageLayout
     eyebrow={t(locale, "papersIndexEyebrow")}
     title={t(locale, "papersIndexTitle")}
     intro={t(locale, "papersIndexIntro")}
   >
     <PapersSection papers={papers} locale={locale} />
-  </NewsLayout>
+  </PageLayout>
 </Layout>
